refactor: #179 로직과 스타일 분리
