+++ src/pages/PostPage/PostCardList/style.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const PostCardListContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  margin: 13px 10px 0 10px;
+`;
