+++ src/pages/PostPage/TagList/style.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const TagListContainer = styled.div`
+  display: flex;
+  gap: 5px;
+  position: fixed;
+  margin: 9px 10px 5px 10px;
+  z-index: 777;
+`;
